requirements.txt
streamlit>=1.30
pandas>=2.0
openpyxl>=3.1
python-docx>=1.1
pdfplumber>=0.10
PyMuPDF>=1.23
Pillow>=10.0
tqdm>=4.66
pytesseract>=0.3.10
